W-064 - move CloseIcon and PencilIcon to icons-folder
