+++ frontend/src/components/user-profile/icons-folder/CloseIcon.jsx
